+++ engine/server/engine/centralized_logs/client_implementations/persistent_volume/persistent_volume_logs_database_client_test.go
@@ -721,11 +721,13 @@ func executeStreamCallAndGetReceivedServiceLogLines(
 		case <-time.Tick(testTimeOut):
 			return nil, stacktrace.NewError("Receiving stream logs in the test has reached the '%v' time out", testTimeOut)
 		case streamErr, isChanOpen := <-errChan:
-			if !isChanOpen {
+			if !isChanOpen && len(userServiceLogsByUuidChan) == 0 {
 				shouldReceiveStream = false
 				break
 			}
-			return nil, stacktrace.Propagate(streamErr, "Receiving streaming error.")
+			if isChanOpen && streamErr != nil {
+				return nil, stacktrace.Propagate(streamErr, "Receiving streaming error.")
+			}
 		case userServiceLogsByUuid, isChanOpen := <-userServiceLogsByUuidChan:
 			if !isChanOpen {
 				shouldReceiveStream = false